Moved TextMessage & Message
