+++ src/guis/krow/guis/chatroom/ChatRoomServer.java
@@ -11,8 +11,8 @@ import krow.guis.chatroom.messages.ChatRoomMessage;
 import krow.guis.chatroom.messages.CommandMessage;
 import kröw.connections.Client;
 import kröw.connections.FullClientListener;
-import kröw.connections.Message;
 import kröw.connections.Server;
+import kröw.connections.messages.Message;
 import kröw.core.Kröw;
 
 public class ChatRoomServer extends Server {
